+++ src/Oro/Bundle/CommerceMenuBundle/Migrations/Schema/OroCommerceMenuBundleInstaller.php
@@ -73,7 +73,7 @@ class OroCommerceMenuBundleInstaller implements
         $table->addColumn('priority', 'integer', ['notnull' => false]);
         $table->addColumn('scope_id', 'integer', ['notnull' => true]);
         $table->addColumn('condition', 'string', ['length' => 512, 'notnull' => false]);
-        $table->addColumn('screens', 'array', ['comment' => '(DC2Type:array)']);
+        $table->addColumn('screens', 'array', ['notnull' => false, 'comment' => '(DC2Type:array)']);
         $table->setPrimaryKey(['id']);
         $table->addUniqueIndex(['key', 'scope_id', 'menu'], 'oro_commerce_menu_upd_uidx');
     }